% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-logo.R
\name{add_correlaid_logo}
\alias{add_correlaid_logo}
\title{Add CorrelAid logo}
\usage{
add_correlaid_logo()
}
\description{
Inset CorrelAid logo to the bottom right corner of a ggplot.
}
\details{
The size of the logo is determined by the base text size set e.g. by
the \code{base_size} argument in \verb{theme_*()} functions.

Adding the logo to a plot makes it a \code{\link[=patchwork-package]{patchwork}}.
Please refer to the patchwork \link[=patchwork-package]{documentation} for
instructions on how to customise its appearance, e.g. via a theme or
scale function.
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(wt, mpg)) + geom_point() + add_correlaid_logo()

ggplot(example_projects_labels, aes(category)) +
 geom_bar() +
 labs(title = "Title", subtitle = "Subtitle", caption = "Caption") +
 theme_correlaid() +
 add_correlaid_logo()

# Change colour scale after adding logo:
ggplot(mtcars, aes(wt, mpg)) +
  geom_point(size = 3) +
  theme_correlaid(base_size = 20) +
  add_correlaid_logo() &
  scale_colour_correlaid_c()
}
