% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlcloud.R
\name{add_correlcloud_user_to_group}
\alias{add_correlcloud_user_to_group}
\title{adds user to group}
\usage{
add_correlcloud_user_to_group(con, user_id, group_id)
}
\arguments{
\item{con}{ocs4R::ocsManager. connection to CorrelCloud as created by new_correlcloud_con}

\item{user_id}{character. user id of the user}

\item{group_id}{character. id of the group the user should be added to.}
}
\value{
boolean. True if successful.
}
\description{
adds user to group
}
