% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\name{add_logo}
\alias{add_logo}
\title{Add CorrelAid logo}
\usage{
add_logo(base_size = 14)
}
\arguments{
\item{base_size}{\code{numeric} Base font size, given in pts}
}
\value{
A \code{inset_path} object
}
\description{
Inset CorrelAid logo to the bottom right corner of a ggplot
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  labs(caption = "") +
  theme_minimal() +
  add_logo(base_size = 11.5)
}
