% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{correlaid_colours}
\alias{correlaid_colours}
\alias{correlaid_colors}
\title{CorrelAid colour palettes}
\format{
An object of class \code{list} of length 3.


}
\usage{
correlaid_colours

correlaid_colors
}
\description{
A list of colour palettes based on the \href{https://docs.correlaid.org/wiki/design-guide}{CorrelAid design guide}
}
\details{
A named list consisting of three character vectors:
\itemize{
\item A "qualitative" palette
\item Two gradient palettes based on the CorrelAid and CorrelAidX design respectively
}
}
\seealso{
\code{\link[=correlaid_pal]{correlaid_pal()}}, \code{\link[=scale_colour_correlaid_d]{scale_colour_correlaid_d()}} etc.
}
\keyword{datasets}
