% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scales.R
\name{correlaid_pal}
\alias{correlaid_pal}
\title{CorrelAid palette}
\usage{
correlaid_pal(direction = 1, option = "qualitative")
}
\arguments{
\item{direction}{Sets the order of colours in the scale:
\itemize{
\item 1: default order
\item -1: reversed order
}}

\item{option}{A character string indicating the colour palette to use.
Three options are available: "qualitative", "gradient" and "gradient_x".
See \code{\link{correlaid_colours}} for more information.}
}
\value{
A function that takes an integer argument (the required number of colours)
and returns a character vector of colours
}
\description{
Create a colour palette based on the \href{https://docs.correlaid.org/wiki/design-guide}{CorrelAid design guide}
}
\examples{
scales::show_col(correlaid_pal()(6))
scales::show_col(correlaid_pal(option = "gradient")(9))
scales::show_col(correlaid_pal(direction = -1, option = "gradient_x")(9))
}
\seealso{
\code{\link{correlaid_colours}}, \code{\link[=scale_colour_correlaid_d]{scale_colour_correlaid_d()}} etc.
}
