% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlaidx-map.R
\name{correlaidx_map}
\alias{correlaidx_map}
\title{correlaidx_map}
\usage{
correlaidx_map(chapters_df = correltools::chapters_df, lang = "en")
}
\arguments{
\item{lang}{character. language of the map. either 'en' or 'de'}

\item{cities_df}{tibble. tibble with all local chapters. defaults to correltools::chapters_df}
}
\description{
returns a widget that shows correlaidx chapters on a map. The input data frame needs specific variables for it
to work, check correltools::chapters_df to see an example.
}
