% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlaidx-map.R
\name{crawl_correlaidx_data}
\alias{crawl_correlaidx_data}
\title{crawl_correlaidx_data}
\usage{
crawl_correlaidx_data(lang = "en")
}
\arguments{
\item{lang}{character. language to crawl chapters for. either 'de' or 'en'. defaults to 'en'}
}
\description{
crawls lc-chapters from the CorrelAid website with package rvest
}
