% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus.R
\name{directus_get}
\alias{directus_get}
\title{get data from directus}
\usage{
directus_get(endpoint, query)
}
\arguments{
\item{endpoint}{character. REST endpoint to get. Refer to \href{https://docs.directus.io/reference/introduction.html}{the API reference}.}

\item{query}{list. list of optional global query parameters.}
}
\value{
list. parsed content of httr return object
}
\description{
get data from directus
}
\details{
utility function that calls directus_get_ with the default arguments. If you need more control over arguments, use directus_get_.
}
