% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus.R
\name{directus_patch}
\alias{directus_patch}
\title{patch/edit data in directus}
\usage{
directus_patch(endpoint, id, body, token = Sys.getenv("DIRECTUS_TOKEN"))
}
\arguments{
\item{endpoint}{character. REST endpoint to patch. Refer to \href{https://docs.directus.io/reference/introduction.html}{the API reference}.}

\item{id}{integer. id of the element to be patched.}

\item{body}{list. updates to patch. do not include the id of the element.}

\item{token}{character. API token for Directus, defaults to environment variable DIRECTUS_TOKEN.}
}
\value{
httr return object
}
\description{
patch/edit data in directus
}
