% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_projects_labels}
\alias{example_projects_labels}
\title{dump of the project labels from the GitHub projects repository}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 237 rows and 10 columns.
}
\usage{
example_projects_labels
}
\description{
dump of the project labels from the GitHub projects repository
}
\keyword{datasets}
