% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlaidx-map.R
\name{get_correlaidx_data}
\alias{get_correlaidx_data}
\title{get_correlaidx_data}
\usage{
get_correlaidx_data(lang = "en")
}
\arguments{
\item{lang}{character. language to get chapters for. either 'de' or 'en'. defaults to 'en'}
}
\description{
gets data for local chapters by crawling the names from the navbar of the CorrelAid website and then geocoding the cities.
also adds information on when the chapter was founded from the package dataset \code{\link{local_chapters_year_founded}}
}
