% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlcloud.R
\name{get_correlcloud_groups}
\alias{get_correlcloud_groups}
\title{list all groups in CorrelCloud}
\usage{
get_correlcloud_groups(con)
}
\arguments{
\item{con}{ocs4R::ocsManager. connection to CorrelCloud as created by new_correlcloud_con}
}
\value{
character. character vector with all group ids in the CorrelCloud.
}
\description{
list all groups in CorrelCloud
}
