% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlcloud.R
\name{get_correlcloud_users}
\alias{get_correlcloud_users}
\title{list all users in CorrelCloud}
\usage{
get_correlcloud_users(con)
}
\arguments{
\item{con}{ocs4R::ocsManager. connection to CorrelCloud as created by new_correlcloud_con}
}
\value{
character. character vector with all user ids in the CorrelCloud.
}
\description{
list all users in CorrelCloud
}
