% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{html_yeti}
\alias{html_yeti}
\title{html_document yeti theme adapted to CorrelAid CI}
\usage{
html_yeti(toc = TRUE, toc_float = TRUE, ...)
}
\arguments{
\item{toc}{logical. whether or not to include a table of contents. defaults to TRUE}

\item{toc_float}{logical. whether or not the toc should float. defaults to TRUE}

\item{...}{other arguments passed down to rmarkdown::html_document}
}
\description{
html_document yeti theme adapted to CorrelAid CI
}
