% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{local_chapters_year_founded}
\alias{local_chapters_year_founded}
\title{Founding years for CorrelAidX chapters}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{chapter_de}{chapter name in German (cf. navbar on \url{https://correlaid.org})}
\item{chapter_en}{chapter name in English (cf. navbar on \url{https://correlaid.org/en})}
\item{year_founded}{year when the chapter was founded}
}
}
\source{
chapter names from \url{https://correlaid.org} respectively \url{https://correlaid.org/de}
}
\usage{
local_chapters_year_founded
}
\description{
A dataset containing the CorrelAidX chapters and when they were founded.
}
\keyword{datasets}
