% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlcloud.R
\name{new_correlcloud_con}
\alias{new_correlcloud_con}
\title{creates new connection to CorrelCloud}
\usage{
new_correlcloud_con(
  usr = Sys.getenv("NEXTCLOUD_USR"),
  pwd = Sys.getenv("NEXTCLOUD_PWD")
)
}
\arguments{
\item{usr}{character. user to use to connect to CorrelCloud. defaults to Sys.getenv("NEXTCLOUD_USR")}

\item{pwd}{character. password to use to connect to CorrelCloud. defaults to Sys.getenv("NEXTCLOUD_PWD")}
}
\value{
ocs4R::ocsManager object that can be used to interact with the CorrelCloud
}
\description{
creates new connection to CorrelCloud
}
