% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlcloud.R
\name{new_correlcloud_group}
\alias{new_correlcloud_group}
\title{creates new group in CorrelCloud}
\usage{
new_correlcloud_group(con, group_id)
}
\arguments{
\item{con}{ocs4R::ocsManager. connection to CorrelCloud as created by new_correlcloud_con}

\item{group_id}{character. name of the group ("id")}
}
\description{
creates new group in CorrelCloud
}
