% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlcloud.R
\name{new_correlcloud_user}
\alias{new_correlcloud_user}
\title{creates new user in CorrelCloud}
\usage{
new_correlcloud_user(
  con,
  first_name,
  last_name,
  email,
  groups = NA,
  user_id = NA
)
}
\arguments{
\item{con}{ocs4R::ocsManager. connection to CorrelCloud as created by new_correlcloud_con}

\item{first_name}{character. first name of user.}

\item{last_name}{character. last name of user}

\item{email}{character. email address to send invitation to.}

\item{groups}{character. character vector of groups to add user to. defaults to NA.}

\item{user_id}{character. user_id/login. defaults to NA, i.e. creating user name from first_name and last_name.}
}
\value{
the created user as per getUser method of ocs4R::ocsManager
}
\description{
creates new user in CorrelCloud
}
