% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_correl_simple.R
\name{theme_correl_simple}
\alias{theme_correl_simple}
\title{Theme inspired by https://correlaid.org/ website}
\usage{
theme_correl_simple(ax_size = 10, axt_size = 14, pt_size = 14, lg_size = 16)
}
\arguments{
\item{ax_size}{size of the axis values. Defaults to 10}

\item{axt_size}{axis titles size. Defaults to 14}

\item{pt_size}{plot title size. Defaults to 14}

\item{lg_size}{legend title size. Defaults to 16}
}
\description{
Theme inspired by the colors on
\href{correlaid.org}{https://correlaid.org/}.
}
