% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\name{theme_correlaid}
\alias{theme_correlaid}
\title{ggplot theme based on CorrelAid design guide}
\usage{
theme_correlaid(base_size = 14, grid = "XY")
}
\arguments{
\item{base_size}{\code{numeric} Base font size, given in pts}

\item{grid}{\code{character} Panel grid ("none" or a combination of X, x, Y, y)}
}
\value{
An object of class \code{theme}
}
\description{
A theme following the \href{https://docs.correlaid.org/wiki/design-guide}{CorrelAid design guide}.
}
\details{
Uses the font family \href{https://fonts.google.com/specimen/Roboto}{Roboto}.
If Roboto is not installed, the system's default (sans-serif) font is used.
}
\examples{
library(ggplot2)
library(showtext)
showtext_auto() # Necessary for certain graphics devices to display the custom font

ggplot(mtcars, aes(wt, mpg)) + geom_point() + theme_correlaid()
}
