% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-logo.R
\name{add_correlaid_logo}
\alias{add_correlaid_logo}
\title{Add CorrelAid logo}
\usage{
add_correlaid_logo()
}
\description{
Inset CorrelAid logo to the bottom right corner of a ggplot.
}
\details{
The size of the logo is determined by the base text size set e.g. by
the \code{base_size} argument in \verb{theme_*()} functions.

Adding the logo to a plot makes it a \code{\link[=patchwork-package]{patchwork}} object.
Using the \code{+} operator e.g. to change the theme or scales after adding the
logo would therefore have no effect. See example below and the
\link[=patchwork-package]{patchwork documentation} for more details.
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(wt, mpg)) + geom_point() + add_correlaid_logo()

ggplot(example_projects_labels, aes(category)) +
 geom_bar() +
 labs(title = "Title", subtitle = "Subtitle", caption = "Caption") +
 theme_correlaid() +
 add_correlaid_logo()

# Change colour scale after adding logo:
# Plot becomes a `patchwork` object after adding logo.
# Use the `&` operator to modify ggplots in a patchwork.
ggplot(mtcars, aes(wt, mpg)) +
  geom_point(size = 3) +
  theme_correlaid(base_size = 20) +
  add_correlaid_logo() +
  scale_colour_correlaid_c()
}
