% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlaidx-map.R
\name{correlaidx_map}
\alias{correlaidx_map}
\title{correlaidx_map}
\usage{
correlaidx_map(chapters_df, lang = "en")
}
\arguments{
\item{chapters_df}{tibble. tibble with all local chapters. best obtained through get_correlaidx_data()}

\item{lang}{character. language of the map. either 'en' or 'de'}
}
\description{
returns a widget that shows correlaidx chapters on a map.
}
