% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus.R
\name{directus_get}
\alias{directus_get}
\title{get data from directus}
\usage{
directus_get(endpoint, token = Sys.getenv("DIRECTUS_TOKEN"), limit = -1)
}
\arguments{
\item{endpoint}{character. REST endpoint to get. Refer to \href{https://docs.directus.io/reference/introduction.html}{the API reference}.}

\item{token}{character. API token for Directus, defaults to environment variable DIRECTUS_TOKEN.}

\item{limit}{integer. how many objects to return. defaults to -1 meaning all objects.}
}
\value{
list. parsed content of httr return object
}
\description{
get data from directus
}
