% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus.R
\name{directus_get_}
\alias{directus_get_}
\title{low level function to get data from directus}
\usage{
directus_get_(
  endpoint,
  parse = TRUE,
  token = Sys.getenv("DIRECTUS_TOKEN"),
  api_base = "https://cms.correlaid.org/",
  limit = -1,
  query = list()
)
}
\arguments{
\item{endpoint}{character. REST endpoint to get. Refer to \href{https://docs.directus.io/reference/introduction.html}{the API reference}.}

\item{parse}{boolean. whether or not to parse the result. defaults to true.}

\item{token}{character. API token for Directus, defaults to environment variable DIRECTUS_TOKEN.}

\item{api_base}{character. API base URL. defaults to https://cms.correlaid.org.}

\item{limit}{integer. how many objects to return. defaults to -1 meaning all objects.}

\item{query}{list. list of optional global query parameters.}
}
\value{
httr return object if \code{parse = FALSE}. parsed content of httr result if \code{parse = TRUE}.
}
\description{
low level function to get data from directus
}
