% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scales.R
\name{scale_colour_correlaid_d}
\alias{scale_colour_correlaid_d}
\alias{scale_color_correlaid_d}
\alias{scale_fill_correlaid_d}
\alias{scale_colour_correlaid_c}
\alias{scale_color_correlaid_c}
\alias{scale_fill_correlaid_c}
\alias{scale_colour_correlaid_b}
\alias{scale_color_correlaid_b}
\alias{scale_fill_correlaid_b}
\title{CorrelAid colour scales for ggplot2}
\usage{
scale_colour_correlaid_d(direction = 1, option = "qualitative", ...)

scale_color_correlaid_d(direction = 1, option = "qualitative", ...)

scale_fill_correlaid_d(direction = 1, option = "qualitative", ...)

scale_colour_correlaid_c(
  direction = 1,
  option = "gradient",
  guide = "colourbar",
  ...
)

scale_color_correlaid_c(
  direction = 1,
  option = "gradient",
  guide = "colourbar",
  ...
)

scale_fill_correlaid_c(
  direction = 1,
  option = "gradient",
  guide = "colourbar",
  ...
)

scale_colour_correlaid_b(
  direction = 1,
  option = "gradient",
  guide = "coloursteps",
  ...
)

scale_color_correlaid_b(
  direction = 1,
  option = "gradient",
  guide = "coloursteps",
  ...
)

scale_fill_correlaid_b(
  direction = 1,
  option = "gradient",
  guide = "coloursteps",
  ...
)
}
\arguments{
\item{direction}{Sets the order of colours in the scale:
\itemize{
\item 1: default order
\item -1: reversed order
}}

\item{option}{A character string indicating the colour palette to use.
Four options are available:
\itemize{
\item "qualitative"
\item "gradient" and "gradient_x", based on the CorrelAid and CorrelAidX logo respectively
\item "grey"
}}

\item{...}{Other arguments passed on to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}},
\code{\link[ggplot2:continuous_scale]{ggplot2::continuous_scale()}}, or \code{\link[ggplot2:binned_scale]{ggplot2::binned_scale()}}
to control name, limits, breaks, labels etc.}

\item{guide}{A function used to create a guide or its name.
See \code{\link[ggplot2:guides]{ggplot2::guides()}} for more information.}
}
\description{
Discrete, continuous and binned ggplot2 colour and fill scales
based on the \href{https://docs.correlaid.org/wiki/design-guide}{CorrelAid design guide}.
For a list of available palettes and colours see \code{\link[=correlaid_pal]{correlaid_pal()}}.
Change the values of the \code{ggplot2.continuous.colour} and \code{ggplot2.continuous.fill}
\code{\link[=options]{options()}} to set these scales as default in ggplot2 (see Details).
}
\details{
Function naming scheme:
\itemize{
\item The suffix \verb{_d} indicates that the scale is discrete.
\item \verb{_c}: continuous scale.
\item \verb{_b}: binned scale.
}

Change the following \code{\link[=options]{options()}} to set these scales as the default ggplot2 scales:

\if{html}{\out{<div class="sourceCode">}}\preformatted{options(
  ggplot2.discrete.colour = correltools::scale_colour_correlaid_d,
  ggplot2.discrete.fill = correltools::scale_fill_correlaid_d,
  ggplot2.continuous.colour = correltools::scale_colour_correlaid_c,
  ggplot2.continuous.fill = correltools::scale_fill_correlaid_c
)
}\if{html}{\out{</div>}}
}
\examples{

library(ggplot2)

ggplot(example_projects_labels, aes(category, fill = category)) +
  geom_bar(show.legend = FALSE) +
  scale_fill_correlaid_d(option = "gradient")

ggplot(mtcars, aes(wt, mpg, colour = mpg)) +
  geom_point(size = 5) +
  scale_colour_correlaid_c()

ggplot(mtcars, aes(wt, mpg, fill = mpg)) +
  geom_point(shape = 21, colour = "white", size = 5) +
  scale_fill_correlaid_b()
}
\seealso{
\code{\link[=correlaid_pal]{correlaid_pal()}}
}
