% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\name{theme_correlaid}
\alias{theme_correlaid}
\title{CorrelAid theme for ggplot2}
\usage{
theme_correlaid(
  base_size = 14,
  base_line_size = base_size/28,
  base_rect_size = base_size/28,
  grid = "XY"
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{grid}{A string: panel grid ("none" or a combination of X, x, Y, y)}
}
\value{
A \code{theme} object
}
\description{
A theme based on the \href{https://docs.correlaid.org/wiki/design-guide}{CorrelAid design guide}
}
\details{
Uses the font family \href{https://fonts.google.com/specimen/Roboto}{Roboto}.
If Roboto is not installed, the system's default (sans-serif) font is used.
}
\examples{
library(ggplot2)
library(showtext)
showtext_auto() # Necessary for certain graphics devices to display the custom font

ggplot(mtcars, aes(wt, mpg)) + geom_point() + theme_correlaid()

ggplot(example_projects_labels, aes(category)) +
 geom_bar() +
 labs(title = "Title", subtitle = "Subtitle", caption = "Caption") +
 theme_correlaid() +
 add_correlaid_logo()
}
